requirements.txt
numpy == 1.26.4
opencv-python == 4.10.0.84
Pillow == 11.0.0
torch == 2.5.1
torchvision == 0.20.1
tqdm == 4.67.1
pydicom == 3.0.1
fastapi == 0.115.6
uvicorn == 0.34.0
scikit-image == 0.25.0
python-multipart == 0.0.20
huggingface-hub == 0.25.2